reinit

+++ python_scripts/make_dirs_and_images.py
@@ -9,15 +9,14 @@ sys.path.append(path.split(path.split(path.abspath(__file__))[0])[0])
 from python_scripts import gen_images, gen_results
 from python_scripts.experiment import Experiment
 
-c_path = sys.argv[1]
-experiments_path = sys.argv[2]
-new_dir_name = datetime.strftime(datetime.now(), "%y.%m.%d-%H:%M:%S")
-
 print("Replacing files...")
-os.mkdir(path.join(experiments_path, new_dir_name))
-shutil.move(path.join(c_path, 'data.bin'), path.join(experiments_path, new_dir_name, 'data.bin'))
-os.chdir(path.join(experiments_path, new_dir_name))
-os.chmod('data.bin', stat.S_IREAD)
+c_path = sys.argv[1]
+if path.exists(path.join(c_path, 'data.bin')):
+    experiments_path = sys.argv[2]
+    new_dir_name = datetime.strftime(datetime.now(), "%y.%m.%d-%H:%M:%S")
+    os.mkdir(path.join(experiments_path, new_dir_name))
+    shutil.move(path.join(c_path, 'data.bin'), path.join(experiments_path, new_dir_name, 'data.bin'))
+    os.chdir(path.join(experiments_path, new_dir_name))
 
 print("Importing data...")
 experiment = Experiment('.')
